SentimentDetection/requirements.txt
torch==2.3.1
transformers==4.49.0
pandas==2.2.3
pillow==11.0.0
torchvision==0.18.0